Move SaltCommand implementations to internal
