+++ modules/apps/portal-crypto/portal-crypto-hash-api/src/main/java/com/liferay/portal/crypto/hash/request/command/salt/SaltCommand.java
@@ -21,5 +21,4 @@ import org.osgi.annotation.versioning.ProviderType;
  */
 @ProviderType
 public interface SaltCommand {
-
 }
